+++ pypsa/statistics/expressions.py
@@ -736,6 +736,9 @@ class StatisticsAccessor(AbstractStatisticsAccessor):
             drop_zero=drop_zero,
             round=round,
         )
+        if self.is_multi_indexed and isinstance(df, pd.DataFrame):
+            weights = self._n.investment_period_weightings["objective"]
+            df = df.multiply(weights, level="periods")
         df.attrs["name"] = "Capital Expenditure"
         df.attrs["unit"] = "currency"
         return df
@@ -847,6 +850,9 @@ class StatisticsAccessor(AbstractStatisticsAccessor):
             drop_zero=drop_zero,
             round=round,
         )
+        if self.is_multi_indexed and isinstance(df, pd.DataFrame):
+            weights = self._n.investment_period_weightings["objective"]
+            df = df.multiply(weights)
         df.attrs["name"] = "Capital Expenditure Fixed"
         df.attrs["unit"] = "currency"
         return df
@@ -1659,6 +1665,9 @@ class StatisticsAccessor(AbstractStatisticsAccessor):
             drop_zero=drop_zero,
             round=round,
         )
+        if self.is_multi_indexed and isinstance(df, pd.DataFrame):
+            weights = self._n.investment_period_weightings["objective"]
+            df = df.multiply(weights)
         df.attrs["name"] = "Operational Expenditure"
         df.attrs["unit"] = "currency"
         return df